// src/lib/api/schemas/userSortByParam.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export type UserSortByParam = typeof UserSortByParam[keyof typeof UserSortByParam];


export const UserSortByParam = {
  display_name: 'display_name',
  created_at: 'created_at',
} as const;