// src/lib/api/schemas/listUsersRequestParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { RoleId } from './roleId.ts';
import type { SortOrderParam } from './sortOrderParam.ts';
import type { UserSortByParam } from './userSortByParam.ts';
import type { UserViewParam } from './userViewParam.ts';

export type ListUsersRequestParams = {
view: UserViewParam;
search?: string;
role?: RoleId;
/**
 * @minimum 0
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
sort_by?: UserSortByParam;
sort_order?: SortOrderParam;
};